tests/requirements-test.txt
pytest
kafka-python
requests
mysql-connector-python
elasticsearch<8.0.0
